requirements.txt
numpy
onnxruntime
huggingface_hub
flask
python-dotenv
